+++ src/app/hi/page.tsx
@@ -21,7 +21,7 @@ export default function RenderPage() {
         >
           <strong>Kizen</strong>
         </Link>
-        . I love all things React, NextJS, and performant web!
+        . I love all things React, Next.js, and performant web!
       </Paragraph>
       <Paragraph className={styles.paragraph}>
         I&apos;d love to connect on{" "}
