+++ src/app/resume/resume.mdx
@@ -11,7 +11,7 @@ import {
 
 Software developer with 9 years of experience with JavaScript and Typescript
 development for the web using modern software stacks, including but not limited
-to NextJS, React, Redux, CI/CD platforms, Amazon Web Services, UI testing, and Google
+to Next.js, React, Redux, CI/CD platforms, Amazon Web Services, UI testing, and Google
 Cloud. Design experience includes Adobe Photoshop, Illustrator, and Figma. A
 lifelong learner always looking for the next way to
 [improve web performance and usability](/blog/building-a-static-generated-blog-with-next-js).
