+++ src/post-word-counts.json
@@ -31,7 +31,7 @@
   "react-interview-questions": 1782,
   "reacting-to-prop-changes-in-react-functional-components": 840,
   "recursive-react-component-rendering": 1874,
-  "reflecting-architecture-decisions-1-year": 2227,
+  "reflecting-architecture-decisions-1-year": 2235,
   "review-your-own-pull-requests": 557,
   "statically-generating-twitter-embeds-and-sharing-for-a-performance-boost": 1329,
   "the-downside-of-optional-chaining-overuse": 1311,
