+++ src/posts/2023-in-review.mdx
@@ -35,7 +35,7 @@ In keeping with what is seemingly an annual tradition, I also wrote about [conne
 
 <PostPreview slug="filling-in-the-connected-fitness-gaps-with-apple-shortcuts" />
 
-In May, I redesigned the entire site and wrote a bit about [the new NextJS features I'm leveraging](https://keegan.codes/blog/welcome-to-my-new-blog):
+In May, I redesigned the entire site and wrote a bit about [the new Next.js features I'm leveraging](https://keegan.codes/blog/welcome-to-my-new-blog):
 
 <PostPreview slug="welcome-to-my-new-blog" />
