+++ src/posts/building-a-static-generated-blog-with-next-js.mdx
@@ -1,4 +1,4 @@
-export const title = "Building a Static Generated Blog with NextJS";
+export const title = "Building a Static Generated Blog with Next.js";
 export const slug = "building-a-static-generated-blog-with-next-js";
 export const tags = ["software", "frontend", "nextjs"];
 export const description =
@@ -16,7 +16,7 @@ When I set out to rebuild my personal website, I had 2 criteria - it needed to b
 
 ## Javascript Frameworks
 
-This website is built using [Next.js](https://nextjs.org/ "NextJS ") and hosted on [Vercel](https://vercel.com/ "Vercel"). I've been using Next.js since its inception for a number of reasons, but the biggest is the ease with which I can set up statically generated pages - meaning each blog post of mine is pre-generated, and has 0 api calls on the client side to fetch the data.
+This website is built using [Next.js](https://nextjs.org/ "Next.js ") and hosted on [Vercel](https://vercel.com/ "Vercel"). I've been using Next.js since its inception for a number of reasons, but the biggest is the ease with which I can set up statically generated pages - meaning each blog post of mine is pre-generated, and has 0 api calls on the client side to fetch the data.
 
 ![](static-blog-post:image-1.png)
