+++ src/posts/statically-generating-twitter-embeds-and-sharing-for-a-performance-boost.mdx
@@ -36,7 +36,7 @@ I really wasn't okay with the performance hit, and showing a few Tweets is simpl
 
 ## Getting the Tweet Data
 
-In order to keep my blog's performance high, I want all the data fetching to happen on the server side when the page is generated. I leverage NextJS's `getStaticProps` for this.
+In order to keep my blog's performance high, I want all the data fetching to happen on the server side when the page is generated. I leverage Next.js's `getStaticProps` for this.
 
 First, I needed to settle on a format for "embedding" Tweets. I settled on a simple HTML tag:
